Refactored email service and auth service
Separated auth concerns from email concerns
