+++ CRMSystemAPI/Services/AuthServices/AuthService.cs
@@ -12,15 +12,13 @@ namespace CRMSystemAPI.Services.AuthServices
     {
         private readonly UserManager<User> _userManager;
         private readonly ITokenService _tokenService;
-        private readonly IEmailSender _emailSender;
         private readonly IConfiguration _configuration;
         private readonly ILogger<AuthService> _logger;
 
-        public AuthService(UserManager<User> userManager, ITokenService tokenService, IEmailSender emailSender, IConfiguration configuration, ILogger<AuthService> logger)
+        public AuthService(UserManager<User> userManager, ITokenService tokenService,IConfiguration configuration, ILogger<AuthService> logger)
         {
             _userManager = userManager;
             _tokenService = tokenService;
-            _emailSender = emailSender;
             _configuration = configuration;
             _logger = logger;
         }
@@ -43,34 +41,6 @@ namespace CRMSystemAPI.Services.AuthServices
                 };
             }
             return null;
-        }
-
-        public async Task SendWelcomeEmailAsync(User user, string password)
-        {
-            var token = await _userManager.GeneratePasswordResetTokenAsync(user);
-            _logger.LogInformation($"Generated token: {token}");
-            var encodedToken = Uri.EscapeDataString(token);
-            _logger.LogInformation($"Encoded token: {encodedToken}");
-            var baseUrl = _configuration["App:BaseUrl"];
-            var resetLink = $"{baseUrl}/auth/reset-password?token={encodedToken}&email={user.Email}";
-
-            var subject = "Welcome to Our Service";
-            var body = $@"
-            Hi {user.Name},
-            <br /><br />
-            Welcome to our service! Here are your account details:
-            <br /><br />
-            Email: {user.Email}<br />
-            Temporary Password: {password}<br /><br />
-            Please reset your password using the following link:
-            <br />
-            <a href='{resetLink}'>Reset Password</a>
-            <br /><br />
-            Best regards,<br />
-            Your Company Team
-            ";
-
-            await _emailSender.SendEmailAsync(user.Email, subject, body);
         }
 
         public async Task<IdentityResult> ResetPasswordAsync(string email, string token, string newPassword)
