+++ CRMSystemAPI/Services/AuthServices/IAuthService.cs
@@ -7,7 +7,6 @@ namespace CRMSystemAPI.Services.AuthServices
     public interface IAuthService
     {
         Task<AuthResponse?> AuthenticateUserAsync(LoginModel loginModel);
-        Task SendWelcomeEmailAsync(User user, string password);
         Task<IdentityResult> ResetPasswordAsync(string email, string token, string newPassword);
     }
 }
