+++ CRMSystemAPI/Services/EmailServices/IEmailSender.cs
@@ -1,7 +1,10 @@
-﻿namespace CRMSystemAPI.Services.EmailServices
+﻿using CRMSystemAPI.Models.DatabaseModels;
+
+namespace CRMSystemAPI.Services.EmailServices
 {
     public interface IEmailSender
     {
         Task SendEmailAsync(string email, string subject, string message);
+        Task SendWelcomeEmailAsync(User user, string password);
     }
 }
